requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.0
pydantic[email]==1.10.13
sqlalchemy==2.0.30
python-dotenv==1.0.1
scikit-learn==1.3.2
numpy==1.26.4
joblib==1.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1